Clean up gateway logic from Plugwise (#91769)
